Enable go linter for yaml-analysis directory.
