+++ .pre-commit-config.yaml
@@ -3,7 +3,7 @@ repos:
     rev: v1.51.2
     hooks:
       - id: golangci-lint
-        args: [--timeout=5m]
+        entry: bash -c 'cd yaml-analysis && golangci-lint run --timeout=5m'
   - repo: https://github.com/igorshubovych/markdownlint-cli
     rev: v0.34.0
     hooks:
