+++ yaml-analysis/helpers.go
@@ -1,185 +1,14 @@
 package main
 
 import (
-	"encoding/csv"
 	"fmt"
 	"os"
 	"sort"
-	"strconv"
 
 	v1 "k8s.io/api/core/v1"
 	networkingv1 "k8s.io/api/networking/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/kubernetes/scheme"
 )
-
-type portConfig struct {
-	singlePorts int
-	portRanges  int
-}
-
-type peerConfig struct {
-	cidrs        int
-	podSelectors int // =gressRules
-	peerPods     int
-}
-
-type gressRule struct {
-	portConfig
-	peerConfig
-}
-
-func (r *gressRule) print(indent string) {
-	fmt.Printf("%s\tports=[single: %v, ranges: %v], peers=[%+v]\n", indent, r.singlePorts, r.portRanges, r.peerConfig)
-}
-
-func (npc *peerConfig) join(npc2 *peerConfig) *peerConfig {
-	if npc2 == nil {
-		return npc
-	}
-	npc.cidrs += npc2.cidrs
-	npc.podSelectors += npc2.podSelectors
-	npc.peerPods = maxInt(npc.peerPods, npc2.peerPods)
-	return npc
-}
-
-type netpolConfig struct {
-	// TODO: diff ingress and egress?
-	localPods  int
-	gressRules []*gressRule
-}
-
-func (c *netpolConfig) print(indent string) {
-	fmt.Printf("%sconfig: localpods=%v, rules:\n", indent, c.localPods)
-	for _, peer := range c.gressRules {
-		peer.print(indent)
-	}
-}
-
-type profileMatch struct {
-	idx    int
-	copies int
-	// sum weight for copies
-	weight float64
-}
-
-type profilesMatch []*profileMatch
-
-func (matches profilesMatch) print(indent string) {
-	fmt.Printf("%smatched profiles:\n", indent)
-	for _, match := range matches {
-		readableMatch := *match
-		readableMatch.idx += 1
-		fmt.Printf("%s\t%+v\n", indent, readableMatch)
-	}
-}
-
-func (matches profilesMatch) weight() float64 {
-	res := 0.0
-	for _, match := range matches {
-		res += match.weight
-	}
-	return res
-}
-
-type netpolWeight struct {
-	npConfig   *netpolConfig
-	result     profilesMatch
-	weight     float64
-	netpolName string
-}
-
-func (w *netpolWeight) print() {
-	fmt.Printf("%v\n", w.netpolName)
-	indent := "  "
-	w.npConfig.print(indent)
-	w.result.print(indent)
-	fmt.Printf("%sweight: %v\n", indent, w.weight)
-}
-
-// <LOCAL_PODS>-<SINGLE_PORTS>-<PORT_RANGES>-<POD_SELECTORS>-<PEER_NAMESPACES>-<PEER_PODS>-<CIDRS>
-type perfProfile struct {
-	localPods    int
-	singlePorts  int
-	portRanges   int
-	podSelectors int
-	// peer namespace just affects the number of peer pods in the end
-	// now used for now
-	//peerNamespaces int
-	peerPods int
-	CIDRs    int
-	// weight = 1/number of policies with this profile
-	weight float64
-}
-
-func newProfile(localPods, gressRules, singlePorts, portRanges, peerPods, peerNamespaces, CIDRs, amount int) *perfProfile {
-	return &perfProfile{localPods, gressRules, singlePorts, portRanges,
-		peerPods * peerNamespaces, CIDRs, 1.0 / float64(amount)}
-}
-
-type podsCounter func(podSelector *metav1.LabelSelector, namespace string, namespaceSelector *metav1.LabelSelector) int
-
-func countSelected(podsList []*v1.Pod, nsList []*v1.Namespace) func(podSelector *metav1.LabelSelector, namespace string, namespaceSelector *metav1.LabelSelector) int {
-	selectedCounter := map[string]int{}
-	return func(podSelector *metav1.LabelSelector, namespace string, namespaceSelector *metav1.LabelSelector) int {
-		stringSelector := podSelector.String() + namespace + namespaceSelector.String()
-		if result, ok := selectedCounter[stringSelector]; ok {
-			return result
-		}
-		matchPodSelector := func(pod *v1.Pod) bool {
-			if podSelector != nil {
-				sel, err := metav1.LabelSelectorAsSelector(podSelector)
-				if err != nil {
-					fmt.Println("ERROR")
-					return false
-				}
-				return sel.Matches(labels.Set(pod.Labels))
-			} else {
-				return true
-			}
-		}
-		matchNamespace := func(ns *v1.Namespace) bool {
-			if namespaceSelector != nil {
-				sel, err := metav1.LabelSelectorAsSelector(namespaceSelector)
-				if err != nil {
-					fmt.Println("ERROR")
-					return false
-				}
-				return sel.Matches(labels.Set(ns.Labels))
-			} else if namespace != "" {
-				return ns.Name == namespace
-			} else {
-				return true
-			}
-		}
-		result := 0
-		matchedNamespaces := sets.Set[string]{}
-		for _, ns := range nsList {
-			if matchNamespace(ns) {
-				matchedNamespaces.Insert(ns.Name)
-			}
-		}
-
-		matchPod := func(pod *v1.Pod) bool {
-			return matchPodSelector(pod) && matchedNamespaces.Has(pod.Namespace)
-		}
-		if len(nsList) == 0 {
-			matchPod = func(pod *v1.Pod) bool {
-				return matchPodSelector(pod) && (namespace == "" || pod.Namespace == namespace)
-			}
-		}
-
-		for _, pod := range podsList {
-			if matchPod(pod) {
-				result += 1
-			}
-		}
-		selectedCounter[stringSelector] = result
-		return result
-	}
-}
 
 func parseYamls(filename string, pods *[]*v1.Pod, namespaces *[]*v1.Namespace, netpols *[]*networkingv1.NetworkPolicy) {
 	content, err := os.ReadFile(filename)
@@ -209,39 +38,6 @@ func parseYamls(filename string, pods *[]*v1.Pod, namespaces *[]*v1.Namespace, n
 			fmt.Printf("WARN: unexpected type %T\n", obj)
 		}
 	}
-}
-
-func parseProfiles(filename string) []*perfProfile {
-	f, err := os.Open(filename)
-	if err != nil {
-		fmt.Printf("ERROR: failed to read file %s: %v\n", filename, err)
-		return nil
-	}
-	defer f.Close()
-	csvReader := csv.NewReader(f)
-	records, err := csvReader.ReadAll()
-	if err != nil {
-		fmt.Printf("ERROR: failed parse profiles: %v\n", err)
-		return nil
-	}
-	profiles := []*perfProfile{}
-	for _, record := range records {
-		ints := []int{}
-		for _, strInt := range record {
-			counter, err := strconv.Atoi(strInt)
-			if err != nil {
-				fmt.Printf("ERROR: failed to convert str %s to int: %v\n", strInt, err)
-				return nil
-			}
-			ints = append(ints, counter)
-		}
-		if len(ints) != 8 {
-			fmt.Printf("ERROR: failed to read a profile: expected 8 ints, got %v\n", len(ints))
-			return nil
-		}
-		profiles = append(profiles, newProfile(ints[0], ints[1], ints[2], ints[3], ints[4], ints[5], ints[6], ints[7]))
-	}
-	return profiles
 }
 
 func maxInt(a, b int) int {
@@ -309,4 +105,17 @@ func topDiv(a, b int) int {
 		res += 1
 	}
 	return res
+}
+
+func ordinalString(i int) string {
+	switch i {
+	case 1:
+		return "1st"
+	case 2:
+		return "2nd"
+	case 3:
+		return "3rd"
+	default:
+		return fmt.Sprintf("%dth", i)
+	}
 }
